Factor out utils from test cases
